Listagem de Inativação

+++ admin/listagem/departamento.php
@@ -10,10 +10,10 @@
 			<div class="card">
 				<div class="card-header">
 					<div class="float-end">
-						<a href="cadastro/departamento" class="btn btn-info">Cadastrar Novo</a>
+						<a href="cadastro/departamento" class="btn btn-primary">Cadastrar Novo</a>
 					</div>
-					<h4>Lista</h4>
-					<h6 class="card-subtitle text-muted">Departamentos</h6>
+					<h4>LISTA</h4>
+					<h6 style="color: green;"><strong>Departamentos Ativos</strong></h6>
 				</div>
 				<table class="table table-bordered table-hover table-striped " id="tabela">
 					<thead>
@@ -33,19 +33,22 @@
 								//separar os dados
 								$id 	    = $dados->id;
 								$nome_dept 	= $dados->nome_dept;
-													
+								$ativo      = $dados->ativo;		
+
 								//mostrar na tela
-								echo '<tr>
-										<td>'.$nome_dept.'</td>
-										<td class="table-action text-center">
-											<a href="cadastro/departamento/'.$id.'" alt="Editar" title="Editar">
-												<i class="align-middle"  data-feather="edit-2"></i>
-											</a>
-											<a href="javascript:excluir('.$id.')" alt="Excluir" title="Excluir">
-												<i class="align-middle" data-feather="trash"></i>
-											</a>
-										</td>
-									</tr>';
+								if ($ativo == "1"){ 
+									echo '<tr>
+											<td>'.$nome_dept.'</td>
+											<td class="table-action text-center">
+												<a href="cadastro/departamento/'.$id.'" alt="Editar" title="Editar">
+													<i class="align-middle"  data-feather="edit-2"></i>
+												</a>
+												<a href="javascript:excluir('.$id.')" alt="Excluir" title="Excluir">
+													<i class="align-middle" data-feather="trash"></i>
+												</a>
+											</td>
+										</tr>';
+								}
 							}
 						?>
 					</tbody>
