+++ admin/listagem/marca.php
@@ -10,10 +10,10 @@
 			<div class="card">
 				<div class="card-header">
 					<div class="float-end">
-						<a href="cadastro/marca" class="btn btn-info">Cadastrar Novo</a>
+						<a href="cadastro/marca" class="btn btn-primary">Cadastrar Novo</a>
 					</div>
-					<h4>Lista</h4>
-					<h6 class="card-subtitle text-muted">Marcas</h6>
+					<h4>LISTA</h4>
+					<h6 style="color: green;"><strong>Marcas Ativas</strong></h6>
 				</div>
 				<table class="table table-bordered table-hover table-striped" id="tabela">
 					<thead>
@@ -32,19 +32,22 @@
 								//separar os dados
 								$id 	        = $dados->id;
 								$nome_marca 	= $dados->nome_marca;
+								$ativo  		= $dados->ativo;
 
-								//mostrar na tela
-								echo '<tr>
-										<td>'.$nome_marca.'</td>
-										<td class="table-action text-center">
-											<a href="cadastro/marca/'.$id.'" alt="Editar" title="Editar">
-												<i class="align-middle"  data-feather="edit-2"></i>
-											</a>
-											<a href="javascript:excluir('.$id.')" alt="Excluir" title="Excluir">
-												<i class="align-middle" data-feather="trash"></i>				
-											</a>
-										</td>
-									</tr>';
+									//mostrar na tela
+								if ($ativo == "1"){
+									echo '<tr>
+											<td>'.$nome_marca.'</td>
+											<td class="table-action text-center">
+												<a href="cadastro/marca/'.$id.'" alt="Editar" title="Editar">
+													<i class="align-middle"  data-feather="edit-2"></i>
+												</a>
+												<a href="javascript:excluir('.$id.')" alt="Excluir" title="Excluir">
+													<i class="align-middle" data-feather="trash"></i>				
+												</a>
+											</td>
+										</tr>';
+								}
 							}
 						?>
 					</tbody>
